organize components into views
